Conditionally load input devices, sync the.view sooner
Handle the lack of a mouse module, for example, by not loading the mouse at all.
It'll cause an error if an app tries to read it... but hopefully the app already
knows what's up.

This also has a fix for the.view not getting updated at startup, if you start it
and immediately unfocus the window (e.g. if running with a console on Windows.)