Handle colliding huge sprites well
Adds special handling for large sprites that would other burn CPU while
spatially hashing. This algorithm is intended to catch a few weirdo sprites,
e.g. a world map colliding with a bunch of small sprites. If lots of sprites
fall under this criteria, it'd be smart to adjust the gridSize of the collision
instead.